+++ backend/src/main/java/com/healthcareapp/backend/entities/PendingQueue.java
@@ -1,4 +1,4 @@
-package com.healthcareapp.backend;
+package com.healthcareapp.backend.entities;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
